fix(audit): cascade preparo 1000x, logout data-leak, modal X dirty-check, BCG decimals
Auditoria funcional (multi-agente + navegador). Correções seguras de alto valor:
- cascadeRecalc: custo_por_kg do preparo agora SEMPRE (custoTotal/rendimento)*1000,
  igual à fórmula canônica do form. Antes só multiplicava p/ g|mL → preparos em
  'un' (e 'ml' minúsculo) ficavam 1000x mais baratos ao atualizar um insumo.
- AuthContext.signOut: limpa clearFeaturesCache() + drafts globais de AsyncStorage
  (entityDraftToRestore/reopenEntityModalAfterEdit/reopenNestedPreparoOnMount) —
  vazavam entre contas em máquina compartilhada (privacidade/integridade).
- EntityCreateModal: botão "X" agora usa handleBackdropPress (dirty-check) em vez
  de onClose direto — evita descartar ficha inteira sem aviso.
- MatrizBCG: handleVendaChange usa parseFloat (parseInt descartava decimais).

+++ src/screens/MatrizBCGScreen.js
@@ -137,9 +137,10 @@ export default function MatrizBCGScreen({ navigation }) {
   }
 
   function handleVendaChange(prodId, value) {
-    // Sessão 28.27: aceita "5", "5,5", "5.5" — parseInt antigo descartava decimais
+    // Sessão 28.27/28.72: aceita "5", "5,5", "5.5". parseInt descartava os decimais
+    // apesar do comentário — trocado por parseFloat (AUDITORIA B3).
     const cleaned = String(value).replace(',', '.').replace(/[^0-9.]/g, '');
-    const qty = parseInt(cleaned) || 0;
+    const qty = parseFloat(cleaned) || 0;
     setVendasMap(prev => ({ ...prev, [prodId]: qty }));
     if (saveTimer.current) clearTimeout(saveTimer.current);
     saveTimer.current = setTimeout(() => saveVenda(prodId, qty), 800);
